Code refactoring - done
