+++ app/src/main/java/com/example/my_test_messenger/MainActivity.kt
@@ -2,24 +2,14 @@ package com.example.my_test_messenger
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.view.View
-import android.widget.Toast
 import androidx.appcompat.widget.Toolbar
 import com.example.my_test_messenger.databinding.ActivityMainBinding
-import com.example.my_test_messenger.ui.ChatFragment
-import com.example.my_test_messenger.ui.SettingsFragment
+import com.example.my_test_messenger.ui.fragments.ChatsFragment
 import com.example.my_test_messenger.ui.objects.AppDrawer
-import com.mikepenz.materialdrawer.AccountHeader
-import com.mikepenz.materialdrawer.AccountHeaderBuilder
-import com.mikepenz.materialdrawer.Drawer
-import com.mikepenz.materialdrawer.DrawerBuilder
-import com.mikepenz.materialdrawer.model.DividerDrawerItem
-import com.mikepenz.materialdrawer.model.PrimaryDrawerItem
-import com.mikepenz.materialdrawer.model.ProfileDrawerItem
-import com.mikepenz.materialdrawer.model.interfaces.IDrawerItem
-import com.mikepenz.materialdrawer.util.DrawerItemViewHelper
-import java.util.Objects
 
+/**
+ * Главный класс приложения, определяет какая тема итд.
+ */
 class MainActivity : AppCompatActivity() {
 
     private lateinit var mBinding: ActivityMainBinding
@@ -46,8 +36,9 @@ class MainActivity : AppCompatActivity() {
          * функциональность и ее выполнение
          */
         setSupportActionBar(mToolbar)
+        mAppDrawer.create()
         supportFragmentManager.beginTransaction()
-            .replace(R.id.dataContainer, ChatFragment()).commit()
+            .replace(R.id.dataContainer, ChatsFragment()).commit()
     }
