+++ app/src/main/java/com/example/my_test_messenger/ui/fragments/ChatsFragment.kt
@@ -1,0 +1,11 @@
+package com.example.my_test_messenger.ui.fragments
+
+import com.example.my_test_messenger.R
+
+class ChatsFragment : BaseFragment(R.layout.fragment_chat) {
+
+    override fun onResume() {
+        super.onResume()
+    }
+
+}
