+++ app/src/main/java/com/example/my_test_messenger/ui/fragments/SettingsFragment.kt
@@ -1,0 +1,11 @@
+package com.example.my_test_messenger.ui.fragments
+
+import com.example.my_test_messenger.R
+
+class SettingsFragment : BaseFragment(R.layout.fragment_settings) {
+
+    override fun onResume() {
+        super.onResume()
+    }
+
+}
